feat: deprecated old tasks and add new ones
